+++ src/ophys_etl/modules/segmentation/postprocess_utils/roi_merging.py
@@ -7,7 +7,7 @@ import copy
 import pathlib
 from ophys_etl.modules.segmentation.postprocess_utils.roi_types import (
     SegmentationROI)
-from ophys_etl.modules.segmentation.postprocess_utils.roi_bayes import (
+from ophys_etl.modules.segmentation.postprocess_utils.roi_time_correlation import (
     validate_merger_corr)
 from ophys_etl.modules.decrosstalk.ophys_plane import get_roi_pixels
 from ophys_etl.modules.decrosstalk.ophys_plane import OphysROI
